Rename Task classes
